requirements.txt
flask
requests
python-dotenv
azure-cognitiveservices-speech